requirements.txt
numpy==1.24.4
opencv-python==4.7.0.72
Pillow==9.4.0
tqdm==4.62.3
matplotlib==3.7.0
matplotlib-inline==0.1.6
